execution/_template.py
"""
Sample execution script template

This is a template for creating deterministic Python scripts in the execution/ folder.
Follow these principles:
- Handle errors gracefully
- Validate inputs
- Log important steps
- Use environment variables for configuration
- Comment your code well
"""

import os
import sys
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def main():
    """
    Main execution function
    """
    try:
        # Your code here
        print("Script executed successfully")
        
    except Exception as e:
        print(f"Error: {str(e)}", file=sys.stderr)
        sys.exit(1)

if __name__ == "__main__":
    main()
