requirements.txt
# ExcelToMif dependencies
# Update with: pip freeze > requirements.txt after every pip install
python-dotenv>=1.0.0
